refactor: re-organize gated command
