+++ snapcraft/commands/validation_sets.py
@@ -22,11 +22,9 @@ import textwrap
 from typing import TYPE_CHECKING
 
 from craft_application.commands import AppCommand
-from craft_cli import emit
-from tabulate import tabulate
 from typing_extensions import override
 
-from snapcraft import const, errors, store
+from snapcraft import const, errors
 
 if TYPE_CHECKING:
     import argparse
@@ -137,59 +135,3 @@ class StoreEditValidationSetsCommand(AppCommand):
             key_name=parsed_args.key_name,
             sequence=parsed_args.sequence,
         )
-
-
-class StoreGatedCommand(AppCommand):
-    """Get snaps and revisions gating a snap."""
-
-    name = "gated"
-    help_msg = "List all gated snaps for <snap-name>"
-    overview = textwrap.dedent(
-        """
-        Get the list of snaps and revisions gating a snap"""
-    )
-
-    @override
-    def fill_parser(self, parser: argparse.ArgumentParser) -> None:
-        parser.add_argument("snap_name", metavar="snap-name")
-
-    @override
-    def run(self, parsed_args: argparse.Namespace):
-        """Print list of snaps gated by snap_name."""
-        store_client = store.StoreClientCLI()
-        account_info = store_client.get_account_info()
-        # Get data for the gating snap
-        snaps = account_info.get("snaps", {})
-
-        # Resolve name to snap-id
-        try:
-            snap_id = snaps[store.constants.DEFAULT_SERIES][parsed_args.snap_name][
-                "snap-id"
-            ]
-        except KeyError:
-            raise store.errors.SnapNotFoundError(snap_name=parsed_args.snap_name)
-
-        validations = store_client.list_validations(snap_id)
-
-        if validations:
-            table_data = []
-            for v in validations:
-                name = v["approved-snap-name"]
-                revision = v["approved-snap-revision"]
-                if revision == "-":
-                    revision = None
-                required = str(v.get("required", True))
-                # Currently timestamps have microseconds, which look bad
-                timestamp = v["timestamp"]
-                if "." in timestamp:
-                    timestamp = timestamp.split(".")[0] + "Z"
-                table_data.append([name, revision, required, timestamp])
-            tabulated = tabulate(
-                table_data,
-                headers=["Name", "Revision", "Required", "Approved"],
-                tablefmt="plain",
-                missingval="-",
-            )
-            emit.message(tabulated)
-        else:
-            emit.message(f"There are no validations for snap {parsed_args.snap_name!r}")
